+++ dlib/binary_search_tree/binary_search_tree_kernel_2.h
@@ -169,7 +169,7 @@ namespace dlib
             );
 
             // functions from the enumerable interface
-            inline unsigned long size (
+            inline size_t size (
             ) const;
 
             bool at_start (
@@ -543,7 +543,7 @@ namespace dlib
         typename mem_manager,
         typename compare
         >
-    unsigned long binary_search_tree_kernel_2<domain,range,mem_manager,compare>::
+    size_t binary_search_tree_kernel_2<domain,range,mem_manager,compare>::
     size (
     ) const
     {
