+++ dlib/interfaces/enumerable.h
@@ -104,7 +104,7 @@ namespace dlib
                 - returns false if there are no more elements in the container
         !*/
 
-        virtual unsigned long size (
+        virtual size_t size (
         ) const = 0;
         /*!
             ensures
